refactor db core
