+++ src/database/core/db-adaptor.ts
@@ -1,0 +1,6 @@
+// import { isDesktop } from '@/const/version';
+import { getDBInstance } from '@/database/core/web-server';
+
+// import { getPgliteInstance } from './electron';
+
+export const serverDB = getDBInstance();
