+++ src/database/repositories/dataImporter/deprecated/__tests__/index.test.ts
@@ -2,6 +2,7 @@
 import { eq, inArray } from 'drizzle-orm/expressions';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
+import { getTestDBInstance } from '@/database/core/dbForTest';
 import {
   agents,
   agentsToSessions,
@@ -11,7 +12,6 @@ import {
   topics,
   users,
 } from '@/database/schemas';
-import { getTestDBInstance } from '@/database/server/core/dbForTest';
 import { CURRENT_CONFIG_VERSION } from '@/migrations';
 import { ImporterEntryData } from '@/types/importer';
